==> 09-Colab-Combining-extracted-covariates-from-subsets.ipynb <==
# Getting started with authentication and package installation

from google.colab import auth
auth.authenticate_user()

!earthengine authenticate
import ee
ee.Initialize()

import subprocess
try:
  import geemap
except ImportError:
  print('geemap package not installed. Installing ...')
  subprocess.check_call(["python", '-m', 'pip', 'install', 'geemap'])

!pip install geopandas

#import geemap
#import os
#import geopandas as gpd
import glob
import pandas as pd
import numpy as np

# Combine biotic and climate covariates files deposited in Google Cloud bucket

!gsutil cp gs://bucket/Moisture/Test/*.csv . # Import biotic and climate covariates extracted for subsets

file_list = []
for files in glob.glob("cov*.csv"):
  file_list.append(files)

result_list = []
for file in file_list:
  result = pd.read_csv(file, usecols = [3,4,5,6,7,8,10,11,12,13,14,15,16,17], header = 0)
  result.columns = ['Coords', 'Date', 'Num', 'Site', 'State', 'Year', 'ppt', 'Tavg', 'VPD', 'GPP', 'EVI', 'NDWI', 'LST', 'Tree']
  siteresult = pd.DataFrame(result)
  result_list.append(siteresult)
  result = pd.DataFrame(np.concatenate(result_list))

result.to_csv('1_final_result.csv') 
!gsutil cp 1_final_result.csv gs://bucket/Moisture/Test/"1_final_result.csv"

# Combine NLDAS files deposited in Google Cloud bucket



!gsutil cp gs://bucket/Moisture/Test2/*.csv . # Import NLDAS extracted for subsets

file_list = []
for files in glob.glob("cov*.csv"):
  file_list.append(files)

result_list = []
for file in file_list:
  result = pd.read_csv(file, usecols = [3,4,5,6,7,8,10,11,12], header = 0)
  result.columns = ['Coords', 'Date', 'Num', 'Site', 'State', 'Year', 'NLDAS1', 'NLDAS2', 'NLDAS3']
  siteresult = pd.DataFrame(result)
  result_list.append(siteresult)
  result = pd.DataFrame(np.concatenate(result_list))

result.to_csv('1_final_result.csv') 
!gsutil cp 1_final_result.csv gs://bucket/Moisture/Test2/"1_final_result.csv"
