==> 14-Colab-Combining-record-QA-flags.ipynb <==
# Getting started with authentication and package installation

from google.colab import auth
auth.authenticate_user()

!earthengine authenticate
import ee
ee.Initialize()

import glob
import pandas as pd
import numpy as np

# Combine quality control files from subsets deposited to the Google Cloud Bucket

# Import quality control files for Modis GPP
# Customize the pathway based on where data is stored here and for following sections
!gsutil cp gs://bucket/Moisture/Test3/failgppset_*.csv .

file_list = []
for files in glob.glob("failgppset*.csv"):
  file_list.append(files)

gppresult_list = []
for file in file_list:
  try: 
    result = pd.read_csv(file, header = 0)
    siteresult = pd.DataFrame(result)
    gppresult_list.append(siteresult)
    gppresult = pd.DataFrame(np.concatenate(gppresult_list))
  except ValueError:
    continue

gppresult = gppresult.iloc[:,[1]]

gppresult.to_csv('1_final_gpp_result.csv') 

!gsutil cp 1_final_gpp_result.csv gs://bucket/Moisture/Test3/"1_final_gpp_result.csv"

!rm *.csv 

# Import quality control files for Modis LST
!gsutil cp gs://rangelands/Moisture/Test3/faillstset_*.csv .

file_list = []
for files in glob.glob("faillstset*.csv"):
  file_list.append(files)

lstresult_list = []
for file in file_list:
  result = pd.read_csv(file, header = 0)
  siteresult = pd.DataFrame(result)
  lstresult_list.append(siteresult)
  lstresult = pd.DataFrame(np.concatenate(lstresult_list))

lstresult = lstresult.iloc[:,[1]]

lstresult.to_csv('1_final_lst_result.csv') 

!gsutil cp 1_final_lst_result.csv gs://bucket/Moisture/Test3/"1_final_lst_result.csv"
